agent-course/unit2/llama_index/requirements.txt
llama-index>=0.9.8
transformers>=4.34.0
torch>=2.0.0
python-dotenv>=1.0.0 